Use doline and don't try to scroll off the screen in redisplay.
While we're at it, window properties width and height (instead of w and h)
so that the window object in w doesn't obliterate the width.

Oh, and we had width and height backwards in the TTYRenderer constructor
arglist.

And don't actually emit the \n in doline, so that we don't (once again) try
to scroll off the sreen.