Make Keymap.interrogate clever about class inheritance hierarchies
The bindings of a method in a child class should override the bindings in
the parent class, both when the method is actually overriden and when the
bindings are on a different method.

That is to say that if I bind 'q' on some method not names self_insert on a
child of editor, that binding is guaranteed to supersede the one on
self_insert, and if I override self_insert, I get whatever bindings there
are on my new method rather than the original method.