bespoke class for chunks
Chunkpocalypse I: The Typening

Begin replacing the list-and-tuple approach with a class so we can play
games with the representation.

One would be tempted to say that this invalidates the "no new types if
possible" approach but deciding that it's time for a class after three
years of experience with it is not giving up on that at all. :-)